refactor: Auto 로그인을 위한 토큰 인증을 UserRepository로 옮기고 GetProfile 연동
issue: #50
