+++ domain/src/main/java/buy/coke/zet/domain/repository/AuthRepository.kt
@@ -5,7 +5,5 @@ import buy.coke.zet.domain.entitiy.login.LoginResponseEntity
 
 interface AuthRepository {
     suspend fun loginWithGoogle(idToken: String): ServiceResult<LoginResponseEntity>
-    suspend fun isValidToken(hasToken: Boolean): ServiceResult<LoginResponseEntity>
-    suspend fun isHasToken(): Boolean
     suspend fun logout(): ServiceResult<Unit>
 }
